add テスト実行できるように改造中。とりあえず存在しないファイル名を入力した場合のテストを追加
